+++ Assets/Scripts/Scene Navigation/GameStarter.cs
@@ -6,10 +6,13 @@ using UnityEngine;
 
 public class GameStarter : MonoBehaviour
 {
+    public GameObject introUI;
 
     public void StartGame()
     {
         SceneManager.LoadScene("Main");
         Time.timeScale = 1f;
+        introUI = GameObject.Find("Intro1");
+        introUI.SetActive(true);
     }
 }
