+++ Assets/Scripts/Scene Navigation/MainLoader.cs
@@ -1,5 +1,3 @@
-//authors: 
-//Alina Polyudova - onCollisionEnter method
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -7,7 +5,7 @@ using UnityEngine.SceneManagement;
 
 public class MainLoader : MonoBehaviour
 {
-
+    // Is called after the garden minigame
     public void LoadMain() {
         SceneManager.LoadScene("Main");
         Time.timeScale = 1f;
