--- Assets/Scripts/UI/QuestDirectionsUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c210dedba0b4f4f44b2ea74e7421275f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
